remove unused connection attribute
fixes https://github.com/edgedb/edgedb-go/issues/126 which didn't have
to do with options.

When running edgedb/edgedb-go against a development server (`edb server`)
the server sent a
[ParameterStatus message](https://www.edgedb.com/docs/internals/protocol/messages#parameterstatus)
that the packaged server builds don't seem to send. The bit of code that
is changed in this PR is not usually exercised because the tests don't
have a way of triggering server initiated messages.