Enable auto decoding JSON values
This change makes it so that any go type can be used as an out value for
EdgeDB json values. `json.Unmarshal` is used under the hood unless the
out value is `[]byte` or `edgedb.OptionalBytes` in which case the
original behavior of passing the json data undecoded is kept.